Upgrade to NanoS+ 1.1.0 and Nano X 2.1.0 (apilevel = 1)

+++ ledger-secure-sdk/nanox/lib_cxng/include/lcx_ecschnorr.h
@@ -32,7 +32,7 @@
 #define LCX_ECSCHNORR_H
 
 /**
- * @brief   Sign a digest message according to the given mode.
+ * @brief   Signs a digest message according to the given mode.
  *
  * @param[in]  pvkey   Pointer to the private key initialized with 
  *                     #cx_ecfp_init_private_key_no_throw beforehand.
@@ -43,6 +43,7 @@
  *                       - CX_ECSCHNORR_BSI03111
  *                       - CX_ECSCHNORR_LIBSECP
  *                       - CX_ECSCHNORR_Z
+ *                       - CX_ECSCHNORR_BIP0340
  *
  * @param[in]  hashID  Message digest algorithm identifier.
  *                     This parameter is mandatory when
@@ -54,7 +55,7 @@
  * @param[in]  msg_len Length of input data.
  *
  * @param[out] sig     ECSchnorr signature encoded in TLV: **30 || L || 02 || Lr || r || 02 || Ls || s**.
- *   
+ *                     This parameter holds the auxiliary random data when CX_ECSCHNORR_BIP0340 is used.
  *
  * @param[in]  sig_len Length of the signature.
  *
@@ -79,10 +80,13 @@ cx_err_t cx_ecschnorr_sign_no_throw(const cx_ecfp_private_key_t *pvkey,
                            size_t *                     sig_len);
 
 /**
- * @brief   Sign a digest message according to the given mode.
+ * @brief   Signs a digest message according to the given mode.
  * 
  * @details This function throws an exception if the computation
  *          doesn't succeed.
+ *
+ * @warning It is recommended to use #cx_ecschnorr_sign_no_throw
+ *          rather than this function.
  *
  * @param[in]  pvkey   Pointer to the private key initialized with 
  *                     #cx_ecfp_init_private_key_no_throw beforehand.
@@ -93,6 +97,7 @@ cx_err_t cx_ecschnorr_sign_no_throw(const cx_ecfp_private_key_t *pvkey,
  *                       - CX_ECSCHNORR_BSI03111
  *                       - CX_ECSCHNORR_LIBSECP
  *                       - CX_ECSCHNORR_Z
+ *                       - CX_ECSCHNORR_BIP0340
  *
  * @param[in]  hashID  Message digest algorithm identifier.
  *                     This parameter is mandatory when
@@ -104,7 +109,7 @@ cx_err_t cx_ecschnorr_sign_no_throw(const cx_ecfp_private_key_t *pvkey,
  * @param[in]  msg_len Length of input data.
  *
  * @param[out] sig     ECSchnorr signature encoded in TLV: **30 || L || 02 || Lr || r || 02 || Ls || s**.
- *   
+ *                     This parameter holds the auxiliary random data when CX_ECSCHNORR_BIP0340 is used.
  *
  * @param[in]  sig_len Length of the signature.
  * 
@@ -122,7 +127,7 @@ cx_err_t cx_ecschnorr_sign_no_throw(const cx_ecfp_private_key_t *pvkey,
  * @throws             CX_EC_INFINITE_POINT
  * @throws             CX_INVALID_PARAMETER_VALUE
  */
-static inline int cx_ecschnorr_sign ( const cx_ecfp_private_key_t * pvkey, int mode, cx_md_t hashID, const unsigned char * msg, unsigned int msg_len, unsigned char * sig, size_t sig_len, unsigned int * info )
+static inline size_t cx_ecschnorr_sign ( const cx_ecfp_private_key_t * pvkey, uint32_t mode, cx_md_t hashID, const unsigned char * msg, unsigned int msg_len, unsigned char * sig, size_t sig_len, unsigned int * info )
 {
   UNUSED(info);
   CX_THROW(cx_ecschnorr_sign_no_throw(pvkey, mode, hashID, msg, msg_len, sig, &sig_len));
@@ -130,7 +135,7 @@ static inline int cx_ecschnorr_sign ( const cx_ecfp_private_key_t * pvkey, int m
 }
 
 /**
- * @brief   Verify a hash message signature according to the given mode.
+ * @brief   Verifies a digest message signature according to the given mode.
  * 
  * @param[in] pukey   Pointer to the public key initialized with 
  *                    #cx_ecfp_init_private_key_no_throw beforehand.
